requirements.txt
streamlit>=1.45.0
pandas==1.5.3
numpy==1.26.0
plotly>=5.24.1
streamlit-option-menu>=0.4.0
streamlit-authenticator>=0.4.2
pyodbc>=5.2.0
PyYAML>=6.0.2
fastapi>=0.115.12
scikit-learn>=1.6.1
statsmodels>=0.14.4
evidently>=0.7.4
geopandas>=1.0.1
matplotlib>=3.10.1
great-expectations>=1.4.2
shapely>=2.1.0
passlib>=1.7.4
python-jose>=3.4.0
requests>=2.32.3
wordcloud>=1.9.4